pyproject.toml
[tool.poetry]
name = "relevator-backend"
version = "0.1.0"
description = ""
authors = ["cmrd-a <cmrd.a@ya.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.64.0"
uvicorn = "^0.13.4"
SQLAlchemy = "^1.4.11"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
python-dotenv = "^0.17.0"
psycopg2-binary = "^2.8.6"
PyJWT = "^2.1.0"
alembic = "^1.6.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.6"
pytest = "^6.2.3"
pytest-cov = "^2.11.1"
requests = "^2.25.1"
mypy = "^0.812"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = "tests"
filterwarnings = [
    'ignore:The explicit passing of coroutine objects:DeprecationWarning',
]